Move `Package` out of `runner`
