+++ README.md
@@ -91,7 +91,7 @@ import asyncio
 
 from yapapi import Executor, Task, WorkContext
 from yapapi.log import enable_default_logger, log_summary, log_event_repr
-from yapapi.runner import vm
+from yapapi.package import vm
 from datetime import timedelta
